feat: moved session Storage to own file
